+++ src/global_vae/training/__init__.py
@@ -1,1 +1,0 @@
-"""training subpackage of global_vae."""